+++ gems/aws-sdk-simpledb/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Fix query serialization with `aws-sdk-core` >= 3.192.0.
+
 1.49.0 (2024-06-25)
 ------------------